Add superflat world generation
